fastapi/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
elastic-transport==8.17.0
elasticsearch==8.17.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
keras==3.9.0
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
milvus-lite==2.4.11
ml-dtypes==0.3.1
mpmath==1.3.0
murmurhash==1.0.12
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pika==1.3.2
pillow==11.1.0
preshed==3.0.9
protobuf==4.25.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymilvus==2.5.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.45.3
sympy==1.13.1
tensorboard==2.16.0
tensorboard-data-server==0.7.2
# tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==2.5.0
tf_keras==2.16.0
tf_keras-nightly==2.16.0.dev2024021410
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
